refactor(image): streamline imports and enhance component structure
