+++ packages/ui/src/components/Image/index.ts
@@ -1,2 +1,3 @@
-export type { ImageProps } from "./Image"
+export type { ImageProps } from "./Image.types"
 export { Image } from "./Image"
+export { shouldShowFallbackImage } from "./Image.utils"
